* Upgraded project to VS2017 * Upgraded boost to 1.65.1 * Remove Xbox solution

+++ OpenSauce/Halo1/Halo1_CE/Interface/OpenSauceUI/GwenUI/c_canvas_gwen.cpp
@@ -82,7 +82,13 @@ namespace Yelo
 			{
 				auto& bounds = m_canvas->GetBounds();
 
-				return rectangle2d { bounds.x, bounds.y, bounds.h, bounds.w };
+				return rectangle2d
+				{
+					static_cast<int16>(bounds.x),
+					static_cast<int16>(bounds.y),
+					static_cast<int16>(bounds.h),
+					static_cast<int16>(bounds.w)
+				};
 			}
 
 			void* c_canvas_gwen::GetControlPtr() const
